requirements.txt
asgiref==3.3.4
Django==3.2.4
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.4.0
django-six==1.0.4
django-widget-tweaks==1.4.8
greenlet==1.1.0
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
screen==1.0.1
six==1.16.0
SQLAlchemy==1.4.17
sqlparse==0.4.1
whitenoise==4.1.3
xlwt==1.3.0
